lambda/pyproject.toml
[project]
name = "lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7,<3.14"
dependencies = [
    "aws-lambda-typing>=2.20.0,<3.0",
    "pillow>=11.3.0,<12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2,<9.0",
]

[tool.pytest.ini_options]
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
